python/requirements.txt
litellm  # for LLM calls
# cryptography  # if needed later
scikit-learn
requests
dilithium-py
python-dotenv
